Add equivocation report logging functionality
Established a new logging function which gets registered as callback to
the equivocation reporter. It creates a new log for each reported
equivocation and filled it with all information of the report, as well
as providing the data to prove this equivocation on a report.
The equivocation report log files use a counter as unique identifier,
since combine proposer and block height only could lead to duplicates.

The block utility functions have been split up, so provide RLP encoding
of block headers standalone next to getting a blocks bare hash.

Related to: #3
See also: trustlines-network/project#299, trustlines-network/project#300