Improve equivocation logging by template
To make the logging of equivocation reports more structured and flush
all information at once to the respective log file, a template has been
introduced. The logging function fetches all information needed for the
template is then instantiating it with a single format instruction,
which gets written to the log.

The approach to use a log counter to solve the problem of unique names
for log files has been reworked, since it will not continue working
after a restart. Instead a single log file per proposer is the new
approach, appending each new equivocation by the same proposer to its
end.

Moreover the equivocation reporter has been extended by a logger, as the
the offline and skip reporter have as well.

Related to: #3
See also: #4