cleanup and unify calls to _insert_branch
set head in _insert_branch, add dedicated method that run the callbacks.

This will help later when storing the state inside the database, because we now
only have one function, that writes to the database.

Please note that we replicated a bug: callbacks for block 0 were not being
called by the block_fetcher (see
https://github.com/trustlines-network/tlbc-monitor/issues/12)