allow to handle session externally from BlockDB object
We now allow to store a persistent session object inside the blockdb. This
allows callers to manage transactions, since BlockDB also will not 'autocommit'
transactions when a persistent session is active.

I wouldn't have written the code in that way in the first place. But completely
removing the BlockDB class requires changes everywhere. And this is the best
thing I could come up with requiring minimal changes.