Check for reorgs based on block before initial one
In case of a reorg based on a block before initial synced one, the block
fetcher would start to sync backwards until the genesis block where he
would fail. Besides this will take a long time where the monitoring
service is not detecting new events, also no meaningful error message
will be shown which is related to the actual problem.
Therefore the block fetcher checks if he is about to going back in
history before the initial block number. In such case an exception will
be thrown and stops the whole monitoring service with related logging.

Related to: #20
See also: trustlines-network/project#356