Add initial block number to persistent state.
To allow detecting of uncontrolled backward syncing due to a reorg, the
initial block number must be known. This variable will become zero on
the first restart of the service. Therefore the variable will be stored
with the state object of the block fetcher to make is persistent over
multiple restarts.

Related to: #20
See also: #21, trustlines-network/project#356