Create validator setup for end-to-end tests
To be able to run end-to-end tests, the configuration files for the
Parity validator nodes are necessary. These includes the client
configuration, the chain specification and the signing/enode keys.

The chain specification is slightly adopted from the default version. It
uses a validator set contract. This contract is provided as pre-compile
within the same file. The source is implementation of
trustlines-protocol/blockchain, but with a initial list of addresses,
since it is not possible to call the init function.

Related to: #37
See also: truslines-network/project#461