Switch to TestValidatorSet contract for e2e tests
To have the opportunity of changing the validator set in the contract
without restrictions for test purpose, the ValidatorSet contract has
been replaced with the TestValidatorSet contract. This contract is used
for unit tests and therefore prepared for this purpose. The initial
address list has not been touched and remains necessary.

Related to: #37
See also: trustlines-network/project#461