Fix immediate transitions for e2e test chain setup
By using a template of Parity the option `immediateTransitions` was set.
For the validator set contract with a history, which is required for the
monitoring service, this must be disabled. The default do so.

Moreover the unnecessary balance for a not used address has been
removed. Also final empty lines or whitespaces have been added where
necessary.

Related to: #37
See also: #42