Add docker-compose setup for e2e tests
To have a setup of multiple containers for a end-to-end test setup with
a running chain and a monitoring service, a docker-compose file has been
written setup it easily.
The existing Dockerfile for the monitor itself has been extended to
provide the default `report` and `db` directories. This removes the need
of any manual directories. At a later point volumes can be easily
mounted.
A new Dockerfile for the Parity client has been added. It is useful to
add the shared configuration files (chain specification, client config
and password) to the image. Docker-compose does not support the
`configs` property for services. Moreover can't multiple Parity
instances share the same volume, since Parity would place its whole
state, database and more into it.

Warning:
The monitoring service can't start correctly at the moment! The patch to
parse validator set contracts is not merged so far.

Related to: #38
See also: trustlines-network/project#461