Add e2e test modules for skip and offline reports
To test if the monitoring service is reporting correctly, two new
modules have been added to the e2e setup. Both modules are based on
Pytest. Tested reports are skipping validators, as well as offline ones.
Both assume that one validator is offline from the beginning.
Furthermore must the monitor command line option for the offline window
be set to `1` to get quick and immediate results.

Related to: #39
See also: trustlines-network/project#461