Change how to calculate offline validators
There were some flaws with the old approach when comparing the missed
steps with the assigned steps, especially if the validator set changed.
We now store the total number of offline steps, which is the number of
the current validators, and then compare that to the window size.
This should also be faster.