Fix the sync message
Before the sync message simply showed the percentage between genisis and
head. Now we store the start sync block number and show the percentage
of the sync from start to head. This is also corrected by the number of
blocks in the current fetched branch. If we get in the near of the
current latest block we will stop showing syncing, and start showing
syncing again if we fall to far behind.