Fix e2e chain specification errors
The following changes have been made by recent tests with a working
monitor with the fixed known bugs.

The monitor is expecting a step duration of 5 seconds, while the
end-to-end chain specification used 1s per default.

Specifying a validator set contract from block `0` has the downside that
the system call to finalize the set does not happen. Therefore the
contracts history will not be initialized as expected by the monitor.
Therefore a transition from a static list is necessary. Tests show that
finality seem to be necessary (Parity output was not clear), so a high
enough transition block number is required.

The pre-compiled validator set contract has initialized the
`currentValidators`. This has worked because of the missing finalization
call. With this call, the `pendingValidators` are written to the
`currentValidators`. Therefore the Parity clients have stopped due to an
empty list of validator addresses. As the `init()` function do, both
list must be initalized equally.

Related to: #39 #45
See also: trustlines-network/project#461