Correct and simply e2e test cases
An offline report is only issued once per validator address. Therefore
the related test cases must be corrected.
Moreover the test cases have been improved by verify that a report file
exist and is explicitly a file only.
The offline window argument for the monitor service was incorrect and
must be the time related to the step duration instead of the slot count.
At least the parsing of the skip and offline report files has been
simplified. This should increase the readability of the test cases.

Related to: #39 #45
See also: trustlines-network/project#461