Fix e2e setup for stable successfull test runs
The problem for the failing tests was due to the connection speed
between the both active validators. If they connect to each other before
they produce the first block, all went fine. But if not, both can seal
a block #1. One node has to reorg to the other one and depending on the
steps, this lead to a detected skip by the reorged valdiator. As fix
a single valdiator is used in the static list of the multi set at the
chain specification. Therefore all the transition to the contract is
safe by finality and the peers have enough time to connect to each
other.

Moreover the `--offline-window` paramter has been adjusted to the
theoretically correct value of one missed step.

At least the test run script was logging two timestamps per docker
container output. Because the Parity clients as well as the monitor use
already their own timestamps for the log messages, this was doubled by
the docker option.\

Related to: #38, #53, #40