Shift validator set transition block
It looks like in the CircleCI environment the transition to the valdiator set
contract was too early. The secondly added valdiator got the signal (not the
appliance) too late and therefore missed his first step sometimes. At least 30
successfull test runs seem to proof this. Before 2 out of 10 failed.

Related to: #60