Simplify report tool for contract interaction
To keep the tool small and better to maintain, the whole setup to get the validator set contract ABI has been simplified. Therefore a fixed and shorten ABI as dictionary has been placed within the core module. In favor all related parts to compile and retrieve the contract including the submodule of the blockchain, have been removed.

Related to: #57
See also: #65