Improve report validator documentation
The connection of the report-validator tool to the according Parity node
was taking a big assumption without explaining it. This has been
extended to point out what the example is assuming and what needs to be
possible adapted. Furthermore an how-to for the quickstart script setup
follows.

Related to: trustlines-protocol/blockchain#445