Wait for node to be fully synced
tlbc-monitor might crash if the node is not fully synced because it
calls into contract that is not synchronized yet.

With this commit the monitor will wait until the node is fully synced.

For some cases this might result in a worse experience. When using
warp-mode, it now also waits until the gap of not-yet synchronized
blocks is gone, whereas previously it was possible to start running
the monitor (though only if the gap was already small enough,
otherwise one could run into the above crash).

If we want to support that case again we need to put in more work and
also look at the block selector stuff, which I assume also doesn't
work with warp-mode.

So, in that sense this is the dumb solution, but it's probably good
enough for now.

Also see https://github.com/trustlines-protocol/tlbc-monitor/issues/83

The node_status module has been copied from the bridge.