Add web3 middleware to handle connection problems
The default middleware of web3 to handle RPC connection problems does
5 attempts directly after each other. This is not helpful for the case
where the node is down for a few seconds or even longer. This is
especially a problem when the node gets automatically updated.
The custom middleware is a fork of the original one. But it introduce
a delay between the retries and do this forever. Additionally the user
gets a log message, stating the current state. This log entry includes the
origin error message. Thereby the user can see what is ongoing.
After all this allows to handle a temporally offline RPC endpoint
without crashing. This solves the problem of updating nodes. Moreover
does it theoretically allow to start the monitor before the node is
reachable.

Related to: #84