Enable web3 retry middleware for all RPC calls
The goals was to stick with the origin http_retry_request middleware as
much as possible. Therefore the custom version did also used RPC method
filter, when to apply the retry functionality or not. Apparently this
list of methods does not fit our needs. The observed problem occurred
already while waiting for the RPC node to be fully synced. The
NodeStatus was using a parity specific call, which is not accepted by
the default filter. Thereby the monitor was still crushing. Without
further concerns, the method filter has been completely removed.

Related to: #84
See also: #87