Improve outdated documentation
This includes the suggestion to use the quickstart setup. Due to this
recommendation and because the README of this `blockchain` repository
include instructions for manual Docker commands, has this section been
removed. This removes overhead and avoids to maintain/synchronize two
places of document the same.

Related to: trustlines-network/project#798